% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{break_help}
\alias{break_help}
\title{Repeating of indexes}
\usage{
break_help(x)
}
\arguments{
\item{x}{Numerical, vector.}
}
\value{
Numerical.
}
\description{
Repeating of indexes
}
\examples{
break_help(c(1, 2, 3))
break_help(c(6, 8, 23, 50))
}
