% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpage_plot.R
\name{ggpage_plot}
\alias{ggpage_plot}
\title{Creates a visualization from the ggpage_build output}
\usage{
ggpage_plot(data, mapping = ggplot2::aes())
}
\arguments{
\item{data}{data.frame. Expects output from \code{ggpage_build} with
optional intermediate analysis.}

\item{mapping}{Default list of aesthetic mappings to use for plot to be
handed to internal \code{ggplot} call.}
}
\value{
A ggplot object with the given visualization.
}
\description{
Creates a visualization from the ggpage_build output
}
\examples{
\dontrun{
library(tidyverse)
library(tidytext)
library(ggpage)
# build and plot
## tibble with full lines
ggpage_build(tinderbox) \%>\%
  ggpage_plot()
## vector with full lines
ggpage_build(book = tinderbox) \%>\%
  pull(text) \%>\%
  ggpage_plot()
## tibble with single words
ggpage_build(tinderbox) \%>\%
  unnest_tokens(text, text) \%>\%
  ggpage_plot()
## vector with single words
ggpage_build(tinderbox) \%>\%
  unnest_tokens(text, text) \%>\%
  pull(text) \%>\%
  ggpage_plot()

# nrow and ncol
ggpage_build(tinderbox, nrow = 2) \%>\%
  ggpage_plot()
ggpage_build(tinderbox, ncol = 2) \%>\%
  ggpage_plot()

# Include analysis within
ggpage_build(tinderbox) \%>\%
  mutate(word_length = str_length(word)) \%>\%
  ggpage_plot(aes(fill = word_length))
}
}
