% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{line_align}
\alias{line_align}
\title{Adjust lines}
\usage{
line_align(line, max_length, type)
}
\arguments{
\item{line}{data.frame}

\item{max_length}{numerical. number of letters allowed on a line.}

\item{type}{Type of line alignment. Must be one of "left", "right" or "both".}
}
\value{
data.frame
}
\description{
Adjust lines
}
