% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_paragraphs.R
\name{nest_paragraphs}
\alias{nest_paragraphs}
\title{converts paragraph tokens into line tokens}
\usage{
nest_paragraphs(data, input, ...)
}
\arguments{
\item{data}{data.frame. With one paragraph per row.}

\item{input}{column that gets split as string or symbol.}

\item{...}{Extra arguments passed to str_wrap.}
}
\value{
data.frame.
}
\description{
extents the str_wrap() function from the stringr package to work with longer
strings.
}
