% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{paper_shape}
\alias{paper_shape}
\title{Identify the egdes of the paper of each page}
\usage{
paper_shape(data)
}
\arguments{
\item{data}{data.frame created by ggpage_build.}
}
\value{
data.frame,
}
\description{
Identify the egdes of the paper of each page
}
\examples{
paper_shape(ggpage_build(tinderbox))
}
