% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{para_index}
\alias{para_index}
\title{paragraph split}
\usage{
para_index(n, FUN, ...)
}
\arguments{
\item{n}{Numeric. Numbers of words.}

\item{FUN}{Numeric. how many words to split whole string by.}

\item{...}{Extra arguments.}
}
\value{
Numeric. paragraph indicator.
}
\description{
Converts a word vector into a line vector with variable paragraph lengths.
}
\details{
FUN most be a function that takes in a number n and returns a vector of
natural numbers.
}
