% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tinderbox}
\alias{tinderbox}
\title{The tinder-box by H.C. Andersen}
\format{A data frame with 211 rows and 2 variables:
\describe{
  \item{text}{character string up to 80 characters each}
  \item{book}{nam of the fairy tale}
  ...
}}
\usage{
tinderbox
}
\description{
A tidy data.frame containing the entire story of The tinder-box by H.C.
Andersen with two columns: \code{text} which contains the text of the
fairy tale divided into elements of up to about 80 characters each and
\code{book} giving the name of the fairy tale in question.
}
\keyword{datasets}
