% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tinderbox_paragraph}
\alias{tinderbox_paragraph}
\title{The tinder-box by H.C. Andersen}
\format{A data frame with 11 rows and 1 variables:
\describe{
  \item{text}{character string up to 80 characters each}
  ...
}}
\usage{
tinderbox_paragraph
}
\description{
A tidy data.frame containing the entire story of The tinder-box by H.C.
Andersen with two columns: \code{text} which contains the text of the
fairy tale divided into paragraphs.
}
\keyword{datasets}
