% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{word_to_line}
\alias{word_to_line}
\title{Internal function for converting words to lines}
\usage{
word_to_line(words, wot_number = 1000)
}
\arguments{
\item{words}{Character. Having each element being seperate words.}

\item{wot_number}{Numeric. how many words to split whole string by.}
}
\value{
Character. have each element be a seperate line.
}
\description{
Internal function for converting words to lines
}
\keyword{internal}
