% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{word_to_line}
\alias{word_to_line}
\title{Internal function for converting words to lines}
\usage{
word_to_line(words, wot_number = 1000)
}
\arguments{
\item{words}{data.frame. Where each row is a separate word words with the
column name text.}

\item{wot_number}{Numeric. how many words to split whole string by.}
}
\value{
Character. have each element be a separate line.
}
\description{
extents the str_wrap() function from the stringr package to work with longer
strings.
}
